+++ src/__tests__/App.test.js
@@ -1,5 +1,6 @@
 import { render, screen } from '@testing-library/react';
-import App from '../components/App';
+
+import App from '../containers/App';
 
 test('renders learn react link', () => {
   render(<App />);
