+++ src/index.js
@@ -1,8 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
-import store from './store';
-import App from './components/App';
+
+import App from './containers/App';
+
+import store from './redux/store';
+
 import './styles/index.scss';
 
 ReactDOM.render(
