+++ src/styles/Results.scss
@@ -1,0 +1,16 @@
+.Results {
+  background-color: #282c34;
+  min-height: 10vh;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: space-between;
+  font-size: calc(10px + 2vmin);
+  color: white;
+  padding: 0 25px;
+
+  input {
+    border-radius: 4px;
+    padding-left: 3px;
+  }
+}
